Refactor responses to wrap in OpenStruct
This makes a bit more palatable for consumers of the API to work with the responses as some top level transformation is done.

It is very light touch on purpose as it is not yet decided if we will have entities, etc.

It also parses the JSON dates (as we are not using ActiveSupport in this gem). This might change if we think is worth the coupling with ActiveSupport