Remove OpenStruct and attributes coupling
We've decided to go with a more abstract/agnostic approach for this http client, meaning we don't really want to "know" anything about what we are receiving from the datastore, we just parse the response as JSON and is up to the individual implementations (Apply and/or Review) to implement decorators, presenters or adapters to this JSON response.

Left just 3 attributes that are supposed to be the bare minimum of an application, but we can even remove these in the future if we see fit. Ideally, we don't want to add any more unless we think is a good idea.