Refreshen code and update dependencies
No functional changes, just enforcing a higher ruby version and updating some outdated gems, like rubocop, faraday, etc.

Fixed a few rubocop offences with hashes shorthand style.

Bumped to v1.0.0 mainly because everything needed for MVP is covered in this version of the gem, and going forward if new endpoints are created, we will add them and start bumping the gem version with it.