CRIMAP-462 Rails engine exposing health endpoints
Basic initial approach to expose the /ping and /health endpoints from the datastore, through Apply and Review by way of using a rails engine that mounts these 2 routes in the host application.

This ultimately behaves similar to a proxy for those endpoints, as datastore does not have an ingress, access is internal to the cluster.

Note despite adding the rails gem to the `Gemfile`, this is only for gem development. At runtime there is no such requirement, so technically this gem is still rails-agnostic and the engine will only load in Rails apps (`if defined?(Rails)`)