CRIMAP-525 Support upload file operation
This endpoint in theory should not be necessary if we are going to use presigned upload URLs but adding it here for completeness, marking as do not merge, can be checkout manually to test or use it, or we can decide we want it. Either way is not very useful until we have an interface.

In order to make this work, we have to support multipart posts, and this is done with a Faraday middleware. A few changes to support this are needed in the `HttpClient` class.