Presigned uploads generate random filename (#12)
We want the filenames to be randomly generated so files are not stored in S3 bucket with their original filenames which means we would have to sanitise some pesky characters and also we could inadvertently leak PII being part of the filenames.

As we are storing each file in its own USN prefix, collisions are almost nonexistent with just a few random chars. Also providers will not be uploading thousand of files per USN anyways.

No changes are required in the Datastore.